move functions to metadata_utils
