+++ matchms/filtering/add_parent_mass.py
@@ -2,7 +2,7 @@ import logging
 from ..constants import PROTON_MASS
 from ..importing import load_adducts_dict
 from ..typing import SpectrumType
-from ..utils import clean_adduct
+from ..metadata_utils import clean_adduct
 
 
 logger = logging.getLogger("matchms")
