+++ matchms/filtering/derive_inchikey_from_inchi.py
@@ -1,8 +1,8 @@
 import logging
 from ..typing import SpectrumType
-from ..utils import convert_inchi_to_inchikey
-from ..utils import is_valid_inchi
-from ..utils import is_valid_inchikey
+from ..metadata_utils import convert_inchi_to_inchikey
+from ..metadata_utils import is_valid_inchi
+from ..metadata_utils import is_valid_inchikey
 
 
 logger = logging.getLogger("matchms")
